+++ src/prob/pf.jl
@@ -202,18 +202,3 @@ function build_mc_pf(pm::AbstractUBFModels)
         constraint_mc_transformer_power(pm, i)
     end
 end
-
-# Deprecated run_ functions (remove in ~4-6 months)
-
-"depreciation warning for `run_ac_mc_pf`"
-function run_ac_mc_pf(data::Union{Dict{String,<:Any},String}, solver; kwargs...)
-    @warn "run_ac_mc_pf is being depreciated in favor of solve_mc_pf(data, ACPUPowerModel, solver; kwargs...), please update your code accordingly"
-    return solve_mc_pf(data, ACPUPowerModel, solver; kwargs...)
-end
-
-
-"depreciation warning for `run_mc_pf`"
-function run_mc_pf(data::Union{Dict{String,<:Any},String}, model_type::Type, solver; kwargs...)
-    @warn "run_mc_pf is being depreciated in favor of solve_mc_pf, please update your code accordingly"
-    return solve_mc_pf(data, model_type, solver; kwargs...)
-end